requirements.txt
Mastodon.py==1.6.1
pymongo==4.3.2
